+++ examples/filter/commonrule/config.json
@@ -1,0 +1,149 @@
+{
+  "servers":[
+     {
+       "default_log_path":"stdout",
+       "listeners":[
+          {
+	    "name":"serverListener",
+	    "address": "127.0.0.1:2046",
+	    "bind_port": true,
+	    "log_path": "stdout",
+	    "filter_chains": [{
+	       "tls_context":{},
+	       "filters": [
+			 {
+	           "type": "proxy",
+		   "config": {
+		      "downstream_protocol": "Http2",
+		      "upstream_protocol": "Http1",
+			 "router_config_name":"server_router"
+
+		   }
+			 },
+			 {
+			 "type":"connection_manager",
+			 "config":{
+			   "router_config_name":"server_router",
+			   "virtual_hosts":[{
+				 "name":"serverHost",
+				 "domains": ["*"],
+				 "routers": [
+				   {
+					 "match":{"prefix":"/"},
+					 "route":{"cluster_name":"serverCluster"}
+				   }
+				 ]
+			   }]
+			  
+			 }
+			 }]
+	    }]
+	  },
+	  {
+	    "name":"clientListener",
+	    "address": "127.0.0.1:2045",
+	    "bind_port": true,
+	    "log_path": "stdout",
+	    "filter_chains": [{
+	       "tls_context":{},
+	       "filters": [
+			 {
+	           "type": "proxy",
+			 "config": {
+				"downstream_protocol": "Http1",
+				"upstream_protocol": "Http2",
+			   "router_config_name":"client_router"
+			 }
+	       },
+			 {
+			   "type":"connection_manager",
+			   "config":{
+				 "router_config_name":"client_router",
+				 "virtual_hosts":[{
+				   "name":"clientHost",
+				   "domains": ["*"],
+				   "routers": [
+					 {
+					   "match":{"prefix":"/"},
+					   "route":{"cluster_name":"clientCluster"}
+					 }
+				   ]
+				 }]
+			   }
+		   }
+		   ]
+	    }],
+        "stream_filters": [
+          {
+            "type": "commonrule",
+            "config": {
+              "ruleConfigs": [
+                {
+                  "enable": true,
+                  "id": 1,
+                  "index": 3,
+                  "limit": {
+                    "limitStrategy": "QPS",
+                    "maxBurstRatio": 1.0,
+                    "periodMs": 1000,
+                    "maxAllows": 5
+                  },
+                  "name": "rule1",
+                  "resources": [
+                    {
+                      "headers": [
+                        {
+                          "compareType": "EQUALS",
+                          "key": "path",
+                          "value": "/serverlist"
+                        },
+                        {
+                          "compareType": "EQUALS",
+                          "key": "method",
+                          "value": "GET"
+                        }
+                      ],
+                      "params": [
+                        {
+                          "compareType": "EQUALS",
+                          "key": "aa",
+                          "value": "a1"
+                        }
+                      ]
+                    }
+                  ],
+                  "runMode": "CONTROL"
+                }
+              ]
+            }
+          }
+        ]
+	  }
+       ]
+     }
+  ],
+  "cluster_manager":{     
+     "clusters":[
+        {
+	  "Name":"serverCluster",
+	  "type": "SIMPLE",
+	  "lb_type": "LB_RANDOM",
+	  "max_request_per_conn": 1024,
+	  "conn_buffer_limit_bytes":32768,
+	  "hosts":[
+	     {"address":"127.0.0.1:8080"}
+	  ]
+	},
+	{
+	  "Name": "clientCluster",
+	  "type": "SIMPLE",
+	  "lb_type": "LB_RANDOM",
+	  "max_request_per_conn": 1024,
+	  "conn_buffer_limit_bytes":32768,
+	  "hosts":[
+	     {"address":"127.0.0.1:2046"}
+	  ]
+	}
+     ]
+  }
+}
